+++ scripts/test/Selenium/Customer/CustomerTicketOverview.t
@@ -12,23 +12,20 @@ use utf8;
 
 use vars (qw($Self));
 
-# get selenium object
 my $Selenium = $Kernel::OM->Get('Kernel::System::UnitTest::Selenium');
 
 $Selenium->RunTest(
     sub {
 
-        # get helper object
         my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
-        # create test customer user
+        # Create test customer user.
         my $TestCustomerUserLogin = $Helper->TestCustomerUserCreate(
         ) || die "Did not get test customer user";
 
-        # get ticket object
         my $TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');
 
-        # create test ticket
+        # Create test ticket.
         my $TicketNumber = $TicketObject->TicketCreateNumber();
         my $TicketID     = $TicketObject->TicketCreate(
             TN           => $TicketNumber,
@@ -51,7 +48,7 @@ $Selenium->RunTest(
         my $InvisibleBody = 'invisible body';
         my $ArticleID     = $TicketObject->ArticleCreate(
             TicketID       => $TicketID,
-            ArticleType    => 'email-internal',
+            ArticleType    => 'note-internal',
             SenderType     => 'agent',
             Subject        => 'an article subject',
             Body           => $InvisibleBody,
@@ -66,14 +63,14 @@ $Selenium->RunTest(
             "ArticleCreate - ID $ArticleID",
         );
 
-        # login test customer user
+        # Login as test customer user.
         $Selenium->Login(
             Type     => 'Customer',
             User     => $TestCustomerUserLogin,
             Password => $TestCustomerUserLogin,
         );
 
-        # search for new created ticket on CustomerTicketOverview screen (default filter is Open)
+        # Search for new created ticket on CustomerTicketOverview screen (default filter is Open).
         $Self->True(
             $Selenium->find_element("//a[contains(\@href, \'Action=CustomerTicketZoom;TicketNumber=$TicketNumber' )]"),
             "Ticket with ticket number $TicketNumber is found on screen with Open filter"
@@ -85,7 +82,34 @@ $Selenium->RunTest(
             'Article body is not visible to customer',
         );
 
-        # check All filter on CustomerTicketOverview screen
+        # Check shown title and article body overview for the test email ticket in the table
+        # for both Ticket::Frontend::CustomerTicketOverview###ColumnHeader settings.
+        for my $ColumnHeader (qw(LastCustomerSubject TicketTitle)) {
+
+            $Helper->ConfigSettingChange(
+                Valid => 1,
+                Key   => 'Ticket::Frontend::CustomerTicketOverview###ColumnHeader',
+                Value => $ColumnHeader,
+            );
+            sleep 1;
+
+            $Self->Is(
+                $Selenium->execute_script(
+                    "return \$('table.Overview tbody tr a[href*=\"Action=CustomerTicketZoom;TicketNumber=$TicketNumber\"]').closest('tr').find('td:contains(\"Untitled!\")').length"
+                ),
+                '1',
+                "Customer Ticket Overview table contains 'Untitled!' as ticket title part",
+            );
+            $Self->Is(
+                $Selenium->execute_script(
+                    "return \$('table.Overview tbody tr a[href*=\"Action=CustomerTicketZoom;TicketNumber=$TicketNumber\"]').closest('tr').find('td:contains(\"This item has no articles yet.\")').length"
+                ),
+                '1',
+                "Customer Ticket Overview table contains 'This item has no articles yet.' as article body part",
+            );
+        }
+
+        # Check All filter on CustomerTicketOverview screen.
         $Selenium->find_element(
             "//a[contains(\@href, \'Action=CustomerTicketOverview;Subaction=MyTickets;Filter=All' )]"
         )->VerifiedClick();
@@ -95,8 +119,8 @@ $Selenium->RunTest(
             "Ticket with ticket number $TicketNumber is found on screen with All filter"
         );
 
-        # check Close filter on CustomerTicketOverview screen
-        # there is only one created ticket, and it should not be on screen with Close filter
+        # Check Close filter on CustomerTicketOverview screen.
+        # There is only one created ticket, and it should not be on screen with Close filter.
         $Selenium->find_element(
             "//a[contains(\@href, \'Action=CustomerTicketOverview;Subaction=MyTickets;Filter=Close' )]"
         )->VerifiedClick();
@@ -106,7 +130,7 @@ $Selenium->RunTest(
             "Ticket with ticket number $TicketNumber is not found on screen with Close filter"
         );
 
-        # clean up test data from the DB
+        # Delete test created ticket.
         my $Success = $TicketObject->TicketDelete(
             TicketID => $TicketID,
             UserID   => 1,
@@ -116,7 +140,7 @@ $Selenium->RunTest(
             "Ticket with ticket number $TicketNumber is deleted"
         );
 
-        # make sure the cache is correct
+        # Make sure the cache is correct.
         $Kernel::OM->Get('Kernel::System::Cache')->CleanUp( Type => 'Ticket' );
 
     }